refactor(sampler): extract shared CPU sampler chain to src/sampler.{h,cpp}
Deduplicate the SamplerCfg / sample_logits / parse_sampler_token block that
test_dflash.cpp and src/laguna_daemon.cpp both carried verbatim. Identical
behaviour: greedy when ` samp=` is absent, otherwise rep_penalty -> top_k ->
softmax(temp) -> top_p -> draw, with the rep_penalty window fed by the
prompt + emitted history. Both translation units now `#include "sampler.h"`
and link against the shared dflash27b lib.

Net -185 LOC across the three files (CMakeLists +1, sampler.{h,cpp} ~150,
removed blocks ~340).

Verified end-to-end on RTX 3090 through scripts/server.py / arch=laguna:
  greedy                 -> "The capital of Spain is Madrid."
  temp=1.5 seed=7        -> "Pluto 🌌"
  temp=1.5 seed=8        -> "Venus."
  daemon banner          -> [test_dflash] arch=laguna -> dispatching to
                            run_laguna_daemon (max_ctx=1024 kv=q8_0
                            chunk=2048). DFlash + DDTree disabled.

No behaviour change on the qwen35 path: same code, just relocated.
